exam_automator/requirements.txt
# Core ProctorIQ Dependencies
# OpenAI for AI-powered evaluation and feedback (optional)
openai>=1.0.0

# Groq API - Fast inference (recommended)
groq>=0.4.0

# Environment variable management
python-dotenv>=1.0.0

# MongoDB - Database for users and submissions
motor>=3.7.0  # Async MongoDB driver for FastAPI
pymongo>=4.15.0

# Authentication & Security
bcrypt>=4.3.0  # Password hashing
pyjwt>=2.10.0  # JWT tokens

# Vector Database - Local FAISS (replaces Pinecone)
faiss-cpu>=1.11.0  # Free local vector similarity search
sentence-transformers>=4.1.0  # Embeddings

# DEPRECATED: Pinecone (keeping for legacy support)
# pinecone-client[grpc]>=3.0.0

# LangChain for document processing and embeddings
langchain>=0.1.0
langchain-community>=0.0.1
# langchain-pinecone>=0.0.1  # DEPRECATED
langchain-huggingface>=0.0.1

# Embeddings and NLP
transformers>=4.30.0
spacy>=3.6.0
nltk>=3.8.0

# Code analysis and quality metrics
pylint>=3.0.0
radon>=6.0.0  # Code complexity metrics
autopep8>=2.0.0
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Style guide enforcement

# Similarity detection and plagiarism checking
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Logging
loguru>=0.7.0

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation and models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File upload and processing
pillow>=10.0.0
pytesseract>=0.3.10
pymupdf>=1.23.0
python-docx>=1.1.0

# Additional dependencies
aiofiles>=23.0.0
httpx>=0.25.0  # For async HTTP requests

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
